cleanup a bit, add question button
