+++ app/components/ExportButton.tsx
@@ -2,137 +2,19 @@ import { useEditor, getSvgAsImage, useToasts, createShapeId } from '@tldraw/tldr
 import { useState } from 'react'
 import { PreviewShape } from '../PreviewShape/PreviewShape'
 import { getHtmlFromOpenAI } from '../lib/getHtmlFromOpenAI'
+import { useMakeReal } from '../hooks/useMakeReal'
 
 export function ExportButton() {
-	const editor = useEditor()
-	const toast = useToasts()
+	const makeReal = useMakeReal()
 
 	// A tailwind styled button that is pinned to the bottom right of the screen
 	return (
 		<button
-			onClick={async (e) => {
-				const newShapeId = createShapeId()
-
-				try {
-					e.preventDefault()
-
-					const selectedShapes = editor.getSelectedShapes()
-
-					if (selectedShapes.length === 0) {
-						toast.addToast({
-							title: 'Nothing selected',
-							description: 'First select something to make real.',
-							id: 'nothing_selected: First select something to make real.',
-						})
-						return
-					}
-
-					const previewPosition = selectedShapes.reduce(
-						(acc, shape) => {
-							const bounds = editor.getShapePageBounds(shape)
-							const right = bounds?.maxX ?? 0
-							const top = bounds?.minY ?? 0
-							return {
-								x: Math.max(acc.x, right),
-								y: Math.min(acc.y, top),
-							}
-						},
-						{ x: 0, y: Infinity }
-					)
-
-					const previousPreviews = selectedShapes.filter((shape) => {
-						return shape.type === 'preview'
-					}) as PreviewShape[]
-
-					if (previousPreviews.length > 1) {
-						toast.addToast({
-							title: 'Too many previous designs',
-							description:
-								'Currently, you can only give the developer one previous design to work with.',
-							id: 'too_many_previous_designs',
-						})
-						return
-					}
-
-					const previousHtml =
-						previousPreviews.length === 1
-							? previousPreviews[0].props.html
-							: 'No previous design has been provided this time.'
-
-					const svg = await editor.getSvg(selectedShapes)
-					if (!svg) {
-						return
-					}
-
-					const IS_SAFARI = /^((?!chrome|android).)*safari/i.test(navigator.userAgent)
-
-					const blob = await getSvgAsImage(svg, IS_SAFARI, {
-						type: 'png',
-						quality: 1,
-						scale: 1,
-					})
-
-					const dataUrl = await blobToBase64(blob!)
-
-					editor.createShape<PreviewShape>({
-						id: newShapeId,
-						type: 'preview',
-						x: previewPosition.x + 60,
-						y: previewPosition.y,
-						props: { html: '', source: dataUrl as string },
-					})
-
-					const json = await getHtmlFromOpenAI({
-						image: dataUrl,
-						html: previousHtml,
-						apiKey:
-							(document.getElementById('openai_key_risky_but_cool') as HTMLInputElement)?.value ??
-							null,
-					})
-
-					if (json.error) {
-						console.error(json)
-						toast.addToast({
-							icon: 'cross-2',
-							title: 'OpenAI API Error',
-							description: `${json.error.message?.slice(0, 100)}...`,
-						})
-						editor.deleteShape(newShapeId)
-						return
-					}
-
-					const message = json.choices[0].message.content
-					const start = message.indexOf('<!DOCTYPE html>')
-					const end = message.indexOf('</html>')
-					const html = message.slice(start, end + '</html>'.length)
-
-					editor.updateShape<PreviewShape>({
-						id: newShapeId,
-						type: 'preview',
-						props: { html, source: dataUrl as string },
-					})
-				} catch (e: any) {
-					console.error(e)
-					toast.addToast({
-						icon: 'cross-2',
-						title: 'Error',
-						description: `Something went wrong: ${e.message.slice(0, 100)}`,
-					})
-					editor.deleteShape(newShapeId)
-				}
-			}}
+			onClick={makeReal}
 			className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded m-2"
 			style={{ cursor: 'pointer', zIndex: 100000, pointerEvents: 'all' }}
 		>
 			Make Real
 		</button>
 	)
-}
-
-export function blobToBase64(blob: Blob): Promise<string> {
-	return new Promise((resolve, _) => {
-		const reader = new FileReader()
-		reader.onloadend = () => resolve(reader.result as string)
-		reader.readAsDataURL(blob)
-	})
 }
