+++ app/globals.css
@@ -34,13 +34,13 @@
 	display: flex;
 	font-size: 12px;
 	font-family: Inter, sans-serif;
-	flex-direction: column;
+	flex-direction: row;
 	max-width: 300px;
 	width: 100%;
-	gap: 8px;
+	gap: 4px;
 	background-color: var(--color-low);
 	border-radius: 8px;
-	padding: 8px;
+	padding: 6px;
 	pointer-events: all;
 }
 
@@ -59,10 +59,12 @@
 	color: transparent;
 	background: var(--color-panel);
 	width: 100%;
+	height: 32px;
 }
 
 .input__wrapper {
 	position: relative;
+	flex-grow: 2;
 }
 
 .input__wrapper:not(:focus-within)::after {
@@ -94,4 +96,17 @@
 
 .lockup__link__mobile {
 	bottom: 60px;
+}
+
+.question__button {
+	all: unset;
+	flex-shrink: 0;
+	width: 32px;
+	height: 32px;
+	background-color: none;
+	border-radius: 4px;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	cursor: pointer;
 }
